+++ winter-junit5/src/test/kotlin/io/jentz/winter/junit5/WinterEachExtensionTest.kt
@@ -35,7 +35,7 @@ class WinterEachExtensionTest {
 
     @BeforeEach
     fun beforeEach() {
-        app.createGraph()
+        app.create()
     }
 
     @Test