+++ winter-rxjava2/src/main/kotlin/io/jentz/winter/rxjava2/WinterDisposablePlugin.kt
@@ -30,9 +30,9 @@ object WinterDisposablePlugin : SimplePlugin() {
 }
 
 fun WinterApplication.installDisposablePlugin() {
-    registerPlugin(WinterDisposablePlugin)
+    plugins += WinterDisposablePlugin
 }
 
 fun WinterApplication.uninstallDisposablePlugin() {
-    unregisterPlugin(WinterDisposablePlugin)
+    plugins -= WinterDisposablePlugin
 }